requirements.txt
fastapi~=0.116.1
uvicorn
transformers~=4.54.1
torch~=2.7.1
chonkie~=1.1.1
qdrant-client
scikit-learn~=1.6.1
pdfplumber~=0.11.7
pytesseract~=0.3.13
pillow~=11.3.0
llama-cpp-python
pydantic~=2.11.7
numpy~=2.0.2